% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAR.R
\name{AAR}
\alias{AAR}
\title{Attribute-wise agreement rate}
\usage{
AAR(x, y)
}
\arguments{
\item{x}{One set of attribute profiles}

\item{y}{The other set of attribute profiles}
}
\value{
The function returns the attribute-wise agreement rate between two sets of attribute profiles.
}
\description{
The function is used to compute the attribute-wise agreement rate between
two sets of attribute profiles. They need to have the same dimensions.
}
