% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestQperm.R
\name{bestQperm}
\alias{bestQperm}
\title{Column permutation of a Q-matrix with respect to a benchmark Q-matrix.}
\usage{
bestQperm(Q, bench.Q)
}
\arguments{
\item{Q}{The targeted Q-matrix.}

\item{bench.Q}{The benchmark Q-matrix.}
}
\value{
The function returns a Q-matrix in which the order of the columns best matches
that of the benchmark Q-matrix.
}
\description{
Function \code{bestQperm} is used to permute the columns of a Q-matrix so that
the order of the columns best matches that of the benchmark Q-matrix. This function
is useful in a Q-matrix estimation process.
}
\examples{
# See examples used for TSQE.

}
