% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAR.R
\name{PAR}
\alias{PAR}
\title{Pattern-wise agreement rate}
\usage{
PAR(x, y)
}
\arguments{
\item{x}{One set of attribute profiles}

\item{y}{The other set of attribute profiles}
}
\value{
The function returns the pattern-wise agreement rate between two sets of attribute profiles.
}
\description{
The function is used to compute the pattern-wise agreement rate between two sets of
attribute profiles. They need to have the same dimensions.
}
