% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distractor.check.R
\name{print.distractor.check}
\alias{print.distractor.check}
\title{Print the Summary of Distractor Check Results}
\usage{
\method{print}{distractor.check}(x, ...)
}
\arguments{
\item{x}{An object of class "distractor.check" returned by \code{\link{distractor.check}}.}

\item{...}{Additional arguments (currently not used).}
}
\description{
Prints a summary of whether the distractors of a given Q-matrix for multiple-choice items are plausible and/or proper.
}
