% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNPC.R
\name{print.GNPC}
\alias{print.GNPC}
\title{Print the Summary of a GNPC Object}
\usage{
\method{print}{GNPC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{GNPC}.}

\item{...}{Additional arguments (not used).}
}
\description{
Prints a summary of the GNPC estimation results.
}
