% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPC.R
\name{print.NPC}
\alias{print.NPC}
\title{Print the Summary of an NPC Object}
\usage{
\method{print}{NPC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"NPC"}.}

\item{...}{Additional arguments (not used).}
}
\description{
Prints a summary of NPC classification results.
}
