% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.completeness.R
\name{print.Qcompleteness}
\alias{print.Qcompleteness}
\title{Print Summary of a Qcompleteness Object}
\usage{
\method{print}{Qcompleteness}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"Qcompleteness"}}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for objects of class \code{"Qcompleteness"}.
}
