% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR.R
\name{print.Qrefine}
\alias{print.Qrefine}
\title{Print Summary of the Q-Matrix Refinement Result}
\usage{
\method{print}{Qrefine}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Qrefine}.}

\item{...}{Additional arguments passed to print methods.}
}
\description{
Prints a summary of the Q-matrix refinement process, including 
which entries were modified and the final refined Q-matrix.
}
