% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.implausible.R
\name{Q.implausible}
\alias{Q.implausible}
\title{Generation of a Q-matrix that contains implausible MC items}
\usage{
Q.implausible(Q, n.implausible)
}
\arguments{
\item{Q}{A proper and plausible Q-matrix for MC items.}

\item{n.implausible}{The number of items that have implausible distractors.}
}
\value{
The function returns
\describe{
\item{Q.implausible}{The generated Q-matrix}
\item{item.implausible}{The ID of the items that have implausible distractors.}
}
}
\description{
This function turns a proper and plausible Q-matrix into an implausible Q-matrix with the 
user-specified number of implausible distractors.
}
\references{
Chiu, C.-Y., Köhn, H. F. & Wang, Y. (Online first). Plausible and proper multiple-choice items 
for diagnostic classification. \emph{Psychometrika}.
\doi{10.1017/psy.2025.10074}
}
