% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.improper.R
\name{Q.improper}
\alias{Q.improper}
\title{Generation of a Q-matrix with improper MC items that are not proper based on a 'plausible and proper' Q}
\usage{
Q.improper(Q, n.improper)
}
\arguments{
\item{Q}{A proper and plausible Q-matrix for MC items.}

\item{n.improper}{The number of improper distractors.}
}
\value{
The function returns
\describe{
\item{Q.improper}{The improper Q-matrix generated from Q}
\item{item.improper}{The ID of the items that are improper.}
}
}
\description{
This function turns a proper and plausible Q-matrix into an improper Q-matrix with the user-specified
number of improper distractors.
}
\references{
Chiu, C.-Y., Köhn, H. F. & Wang, Y. (Online first). Plausible and proper multiple-choice items 
for diagnostic classification. \emph{Psychometrika}.
\doi{10.1017/psy.2025.10074}
}
