% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{RR}
\alias{RR}
\title{Entry-wise and vector-wise agreement rate between two Q-matrices}
\usage{
RR(Q1, Q2)
}
\arguments{
\item{Q1}{The first Q-matrix.}

\item{Q2}{The second Q-matrix that has the same dimensionality as \code{Q1}.}
}
\value{
The function returns
\describe{
\item{entry.wise}{The entry-wise agreement rate}
\item{item.wise}{The item-wise agreement rate}
}
}
\description{
\code{RR} is used to compute the agreement rate between two Q-matrices with 
identical dimensions.
}
\seealso{
See the examples for the \code{\link{QR}} and \code{\link{TSQE}} functions.
}
