% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestQperm.R
\name{bestQperm}
\alias{bestQperm}
\title{Column permutation of the estimated Q-matrix}
\usage{
bestQperm(estQ, trueQ)
}
\arguments{
\item{estQ}{The estimated Q-matrix.}

\item{trueQ}{The true Q-matrix.}
}
\value{
The function returns a Q-matrix in which the order of the columns best matches
that of the true Q-matrix.
}
\description{
Function \code{bestQperm} is used to rearrange the columns of the estimated Q so that
the order of the columns best matches that of the true Q-matrix.
}
