% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_generate.R
\name{Q.generate}
\alias{Q.generate}
\title{Generation of Dichotomous Q-Matrix}
\usage{
Q.generate(K, J, p, single.att = TRUE)
}
\arguments{
\item{K}{The number of attributes}

\item{J}{The number of items}

\item{p}{The probability of getting a one in the Q-matrix}

\item{single.att}{Whether all the single attribute patterns are included.
If \code{T}, the completeness of the Q-matrix is guaranteed.}
}
\value{
The function returns a complete dichotomous Q-matrix
}
\description{
The function generates a complete Q-matrix based on a
pre-specified probability of getting a one.
}
\examples{
q = Q.generate(3,20,0.5,single.att = TRUE)
q1 = Q.generate(5,30,0.6,single.att = FALSE)

}
