% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{RR}
\alias{RR}
\title{Entry-wise and vector-wise recovery rates}
\usage{
RR(Q1, Q2)
}
\arguments{
\item{Q1}{The first Q-matrix.}

\item{Q2}{The second Q-matrix that has the same dimensionality as Q1.}
}
\value{
The function returns
\describe{
\item{entry.wise}{The entry-wise recovery rate}
\item{item.wise}{The item-wise recovery rate}
}
}
\description{
Function \code{RR} is used to compute the recovery rates for an estimate Q-matrix.
In general, it can be used to compute the agreement rate between two matrices with identical dimensionalities.
}
